Skip filtered fields and improve date comparison
A field filter specifies actions on the field value.
The field value in GOB is therefor no longer comparable with the original source value.

Date comparison has been enhanced so that dates can be compared with date-times.
If the datetime has a 00:00:00 time then the datetime value compares with a date value.